boeing-main/requirements.txt
pdfminer_six==20250416
playwright==1.46.0
Requests==2.32.2
openai==1.63.0
chardet==5.2.0
together==1.5.5
pandas==2.2.2
python-dotenv==1.0.1
torch==2.6.0